requirements.txt
vk-api==11.9.9
python-dotenv==0.20.0
yadisk==1.2.15
